Update Enemy
- ~~Fix dead animation~~
- ~~Fix Hide range enemy FUK!~~
- ~~Fix FlyEnemy Animation~~
    - Delete
Fix Warrior Chaos Animation / System

+++ ProjectGameD/Assets/Enemy Asset/EnemyAI/AI ver.3/Enemy Health.cs
@@ -14,7 +14,7 @@ public class EnemyHealth : MonoBehaviour
     private float currentHealth;
     public GameObject dmgtext;
 
-    public void SetState(int IN_Health)
+    public void SetState(float IN_Health)
     {
         Health = IN_Health;
         maxHealth = IN_Health;
@@ -24,7 +24,6 @@ public class EnemyHealth : MonoBehaviour
     public void CalculateDamage(float playerWeaponDamage)
     {
         currentHealth -= playerWeaponDamage;
-
         currentHealth = Mathf.Clamp(currentHealth, 0, maxHealth); // Ensure health doesn't go below 0
     }
